Nolan Hamilton: Merging 'Wood Chop' fix with local.

+++ src/edu/ycp/cs320/IslandAdventure/controller/ActionController.java
@@ -1,5 +1,7 @@
 package edu.ycp.cs320.IslandAdventure.controller;
 
+import edu.ycp.cs320.IslandAdventure.model.*;
+
 public class ActionController 
 {
 	InventoryController inventoryController = new InventoryController();
@@ -8,7 +10,7 @@ public class ActionController
 	{
 		if (action.equals("Chop Wood"))
 		{
-			inventoryController.changeWoodAmount(10.0);
+			inventoryController.changeWoodAmount((double) 10.0);
 		}
 	}
 }
